requirements.txt
numpy
torch
torchvision
tqdm
Pillow